fix to issue where file isn't renamed
DataLogManager waits until there are updates from the driver station, so
if we spam `DriverStationSim.notifyNewData()`, it will be able to update
properly. Also, in order to make the hoot files show up, we need to wait
a bit longer, so a sleep was added.
Is this dumb? Yes. Does it work? Also yes.