Variants of left-{0200,0300} autos with robot reversed at coral station (#158)
This will allow us to experiment with how much better our estimates robot
poses could be if our offseason 2025 robot did the outtake on the opposite
side as the intake.

For both new paths, the robot tries to point towards AprilTag 17 as it moves
towards the coral station.

Created by copying existing paths and making manual changes to change the
rotation for the goal and state, and changing/adding point towards zones.