Write tests for auto routines that use the drivetrain
For some reason, on test runs that are not the first, PathPlanner freaks out and is completely inaccurate. Thus, the auto tests only fail when they aren't run first. I have no clue what is going on. At one point both succeeded, but they just went back to failing; this test is a house of cards.

The forward tests probably aren't even a good test of forward working after auto; that would require dynamically finding an auto routine in the deploy folder that resets your position when run. I don't really want to have a bad time, so I am not going to do that yet